+++ bindings/react/src/components/BackButton.jsx
@@ -1,5 +1,5 @@
 import SimpleWrapper from './SimpleWrapper.jsx';
-import React from 'react';
+import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 
 /**
@@ -52,7 +52,7 @@ BackButton.propTypes = {
    *  [en]The appearance of the back button.[/en]
    *  [jp][/jp]
    */
-  modifier: React.PropTypes.string,
+  modifier: PropTypes.string,
 
   /**
    * @name onClick
@@ -61,7 +61,7 @@ BackButton.propTypes = {
    *  [en]This function will be called ones the button is clicked. It overrides the default behavior of the back button.[/en]
    *  [jp][/jp]
    */
-  onClick: React.PropTypes.func
+  onClick: PropTypes.func
 };
 
 export default BackButton;