+++ bindings/react/src/components/Toolbar.jsx
@@ -1,5 +1,6 @@
 import SimpleWrapper from './SimpleWrapper.jsx';
-import React from 'react';
+
+import PropTypes from 'prop-types';
 
 /**
  * @original ons-toolbar
@@ -44,7 +45,7 @@ Toolbar.propTypes = {
    *  [/en]
    *  [jp] どうしよう[/jp]
    */
-  modifier: React.PropTypes.string
+  modifier: PropTypes.string
 };
 
 export default Toolbar;